Create DrivSafe
We want to create DrivSafe software, which will be connected to the airbags and sudden brake and a sensor on the front bumper of the car so that in case the driver met an accident then immediately there will be a ping on nearest police station, hospital,NGO and he will be given immediate first aid.

Also this can be used to locate and appoint a nearby motor mechanic or service centre in case of disfunctioning of the vehicle. 
This will use AI,IoT and C coding.